tests/wpt/web-platform-tests/offscreen-canvas/the-canvas-state/2d.state.saverestore.shadowOffsetY.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.state.saverestore.shadowOffsetY</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.state.saverestore.shadowOffsetY</h1>
<p class="desc">save()/restore() works for shadowOffsetY</p>


<script>
var t = async_test("save()/restore() works for shadowOffsetY");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

// Test that restore() undoes any modifications
var old = ctx.shadowOffsetY;
ctx.save();
ctx.shadowOffsetY = 5;
ctx.restore();
_assertSame(ctx.shadowOffsetY, old, "ctx.shadowOffsetY", "old");

// Also test that save() doesn't modify the values
ctx.shadowOffsetY = 5;
old = ctx.shadowOffsetY;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against 5
ctx.save();
_assertSame(ctx.shadowOffsetY, old, "ctx.shadowOffsetY", "old");
ctx.restore();
t.done();

});
</script>
